requirements.txt
Django==2.2
asgiref==3.2.5
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
django-rest-swagger==2.2.0
djangorestframework==3.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
openapi-codec==1.3.2
pytz==2019.3
requests==2.24.0
simplejson==3.17.2
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
